Finish factory tests
